pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.2.1",
    "langchain-text-splitters>=0.3.0",
    "markitdown[pdf]>=0.1.3",
    "openai>=2.6.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.50.0",
]
